# backend/docs/swagger.yaml
basePath: /
definitions:
  controller.AddToCollectionRequest:
    properties:
      answer_id:
        type: integer
      collection_id:
        type: integer
    required:
    - answer_id
    - collection_id
    type: object
...
  controller.CreateCollectionRequest:
    properties:
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个类似知乎的问答平台 API 文档
  title: AgentTalk API
  version: "1.0"
paths:
...
  /collection:
    post:
      consumes:
      - application/json
      description: 创建一个新的收藏夹
      parameters:
      - description: 收藏夹名称
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"创建成功\", \"data\": {...}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 创建收藏夹
      tags:
      - 收藏管理
  /collection/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定收藏夹（仅拥有者可操作）
      parameters:
      - description: 收藏夹ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"删除成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"收藏夹不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 删除收藏夹
      tags:
      - 收藏管理
  /collection/item:
    delete:
      consumes:
      - application/json
      description: 从指定收藏夹中移除答案
      parameters:
      - description: 收藏夹ID
        in: query
        name: collection_id
        required: true
        type: integer
      - description: 答案ID
        in: query
        name: answer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"移除成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"收藏夹不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 从收藏夹移除答案
      tags:
      - 收藏管理
    post:
      consumes:
      - application/json
      description: 将答案添加到指定收藏夹
      parameters:
      - description: 收藏参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.AddToCollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"message\": \"收藏成功\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"收藏夹或答案不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "409":
          description: '{\"code\": 409, \"message\": \"已经收藏过了\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 添加答案到收藏夹
      tags:
      - 收藏管理
  /collection/items:
    get:
      consumes:
      - application/json
      description: 获取指定收藏夹中的所有答案，使用游标分页，包含统计信息
      parameters:
      - description: 收藏夹ID
        in: query
        name: collection_id
        required: true
        type: integer
      - description: 游标（上一页最后一条记录的ID）
        in: query
        name: cursor
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": {\"list\": [...], \"next_cursor\":
            123, \"has_more\": true}}'
          schema:
            $ref: '#/definitions/controller.Response'
        "400":
          description: '{\"code\": 400, \"message\": \"参数错误\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "403":
          description: '{\"code\": 403, \"message\": \"无权限操作\"}'
          schema:
            $ref: '#/definitions/controller.Response'
        "404":
          description: '{\"code\": 404, \"message\": \"收藏夹不存在\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 获取收藏夹中的答案列表
      tags:
      - 收藏管理
  /collection/list:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有收藏夹
      produces:
      - application/json
      responses:
        "200":
          description: '{\"code\": 200, \"data\": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
        "401":
          description: '{\"code\": 401, \"message\": \"未授权\"}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - BearerAuth: []
      summary: 获取收藏夹列表
      tags:
      - 收藏管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"